requirements.txt
openai==1.0.0
langchain==0.0.267
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0